Moving router/werker to server/grpc/werker. Renamed the websocket file to include _ws_. It doesn't cleanly belong in this package
